+++ src/domain/datasets/services/src/use_cases/versioned_files/update_version_file_use_case_impl.rs
@@ -19,13 +19,13 @@ use kamu_core::{
     PushIngestDataUseCase,
     PushIngestError,
     QueryService,
-    ResolvedDataset,
     auth,
 };
 use kamu_datasets::{
     ContentArgs,
     ExtraDataFields,
     FileVersion,
+    ResolvedDataset,
     UpdateVersionFileResult,
     UpdateVersionFileUseCase,
     UpdateVersionFileUseCaseError,